requirements.txt
Django==4.2
djangorestframework==3.14.0
psycopg2-binary==2.9.6
celery==5.3.0
redis==4.5.4
django-celery-beat==2.5.0
gunicorn==20.1.0